Initial commit / v 1.0.0
This repository is intended as a convenience package to enable people to easily install all the PHPCS standards in the PHPCompatibility organisation in one go using Composer.

In other words: the only "functionality" which will be contained within this package will be the `composer.json` file.
In due time, more standards may be added to that file.

Notes:
- Adds `composer.json` file which requires all standards currently offered by the PHPCompatibility organisation.
- Adds a Travis config to validate the Composer configuration on each update.
- Adds README with basic installation and usage instructions.
- Adds LGPL v3 License.
- Adds `.gitignore` to ignore the `vendor` directory and the `composer.lock` file.
- Adds `.gitattributes` to keep release archives clean.
- Adds `.github/issue_template.md` to try and prevent issues related to the standards from being opened in this repository.